backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Приложение для групповых денеженыных сборов средств"
authors = ["Anton <anton_ata@mail.ru>"]
readme = "../README.md"

[tool.poetry.dependencies]
python = "^3.10"
amqp = "5.2.0"
asgiref = "3.8.1"
async-timeout = "4.0.3"
attrs = "23.2.0"
billiard = "4.2.0"
celery = "5.4.0"
certifi = "2024.2.2"
cffi = "1.16.0"
charset-normalizer = "3.3.2"
click = "8.1.7"
click-didyoumean = "0.3.1"
click-plugins = "1.1.1"
click-repl = "0.3.0"
coverage = "7.5.1"
cryptography = "42.0.7"
defusedxml = "0.8.0rc2"
django = "4.2.10"
django-templated-mail = "1.1.1"
djangorestframework = "3.14.0"
djangorestframework-simplejwt = "5.3.1"
djoser = "2.2.2"
drf-spectacular = "0.27.2"
factory-boy = "3.3.0"
faker = "25.2.0"
flake8 = "7.0.0"
flake8-isort = "6.1.1"
gunicorn = "22.0.0"
idna = "3.7"
inflection = "0.5.1"
install = "1.3.5"
isort = "5.13.2"
jsonschema = "4.22.0"
jsonschema-specifications = "2023.12.1"
kombu = "5.3.7"
mccabe = "0.7.0"
oauthlib = "3.2.2"
packaging = "24.0"
pillow = "10.3.0"
prompt-toolkit = "3.0.43"
psycopg2-binary = "2.9.9"
pycodestyle = "2.11.1"
pycparser = "2.22"
pyflakes = "3.2.0"
pyjwt = "2.8.0"
python-dateutil = "2.9.0.post0"
python-dotenv = "1.0.1"
python3-openid = "3.2.0"
pytz = "2024.1"
pyyaml = "6.0.1"
redis = "5.0.4"
referencing = "0.35.1"
requests = "2.31.0"
requests-oauthlib = "2.0.0"
rpds-py = "0.18.1"
six = "1.16.0"
social-auth-app-django = "5.4.1"
social-auth-core = "4.5.4"
sqlparse = "0.5.0"
typing-extensions = "4.11.0"
tzdata = "2024.1"
uritemplate = "4.1.1"
urllib3 = "2.2.1"
vine = "5.1.0"
wcwidth = "0.2.13"

[tool.isort]
multi_line_output = 3
combine_as_imports = true
include_trailing_comma = false

force_grid_wrap = 2

use_parentheses = true
ensure_newline_before_comments = true
line_length = 79
indent = 4
atomic = true
case_sensitive = false
balanced_wrapping = false

[tool.poetry.group.dev.dependencies]
setuptools = "^69.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
